+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonDataGet.java
@@ -10,6 +10,7 @@ import pnet.data.api.util.ById;
 import pnet.data.api.util.GetFunction;
 import pnet.data.api.util.IncludeInactive;
 import pnet.data.api.util.Pair;
+import pnet.data.api.util.RestrictCredentialsAvailable;
 import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictEmail;
 import pnet.data.api.util.RestrictExternalId;
@@ -21,10 +22,10 @@ import pnet.data.api.util.RestrictTenant;
 /**
  * @author cet
  */
-public class PersonDataGet extends AbstractGet<PersonDataDTO, PersonDataGet>
-    implements RestrictTenant<PersonDataGet>, RestrictExternalId<PersonDataGet>, RestrictGuid<PersonDataGet>,
-    RestrictPreferredUserId<PersonDataGet>, RestrictEmail<PersonDataGet>, RestrictDatedBackUntil<PersonDataGet>,
-    RestrictPersonnelNumber<PersonDataGet>, IncludeInactive<PersonDataGet>, ById<PersonDataDTO, PersonDataGet>
+public class PersonDataGet extends AbstractGet<PersonDataDTO, PersonDataGet> implements RestrictTenant<PersonDataGet>,
+    RestrictExternalId<PersonDataGet>, RestrictGuid<PersonDataGet>, RestrictPreferredUserId<PersonDataGet>,
+    RestrictEmail<PersonDataGet>, RestrictDatedBackUntil<PersonDataGet>, RestrictPersonnelNumber<PersonDataGet>,
+    RestrictCredentialsAvailable<PersonDataGet>, IncludeInactive<PersonDataGet>, ById<PersonDataDTO, PersonDataGet>
 {
 
     public PersonDataGet(GetFunction<PersonDataDTO> getFunction, List<Pair<String, Object>> restricts)